requirements.txt
accelerate==1.11.0
coverage==7.11.0
datasets==4.4.1
evaluate==0.4.6
flake8==7.3.0
huggingface-hub==0.36.0
mypy==1.14.0
mypy-extensions==1.0.0
numpy==2.1.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
pandas==2.2.3
pip==24.2
pydantic==2.12.4
pydantic_core==2.41.5
pytest==8.3.3
pytest-cov==7.0.0
PyYAML==5.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing_extensions==4.15.0
typing-inspection==0.4.2